+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Cursor agent support for MCP server configuration
 - Qoder agent support for MCP server configuration
 - Support for project-level MCP configurations
+- Filter MCP Server by name
 
 
 ## [0.0.5] - 2025-10-29